+++ node/constants.ts
@@ -9,6 +9,7 @@ export const QUOTE_FIELDS = [
   'id',
   'referenceName',
   'creatorEmail',
+  'creatorName',
   'creatorRole',
   'creationDate',
   'expirationDate',
@@ -23,8 +24,10 @@ export const QUOTE_FIELDS = [
   'viewedByCustomer',
   'salesChannel',
   'seller',
+  'sellerName',
   'parentQuote',
   'hasChildren',
+  'childrenQuantity',
 ]
 
 export const routes = {
@@ -87,6 +90,10 @@ export const schema = {
       title: 'Creator Email',
       type: 'string',
     },
+    creatorName: {
+      title: 'Creator Name',
+      type: 'string',
+    },
     creatorRole: {
       title: 'Creator Role',
       type: 'string',
@@ -141,6 +148,10 @@ export const schema = {
       title: 'Seller',
       type: ['null', 'string'],
     },
+    sellerName: {
+      title: 'Seller Name',
+      type: ['null', 'string'],
+    },
     parentQuote: {
       title: 'Parent quote',
       type: ['null', 'string'],
@@ -149,11 +160,16 @@ export const schema = {
       title: 'Has children',
       type: ['null', 'boolean'],
     },
+    childrenQuantity: {
+      title: 'Children Quantity',
+      type: ['null', 'number'],
+    },
   },
   'v-cache': false,
   'v-default-fields': [
     'referenceName',
     'creatorEmail',
+    'creatorName',
     'creationDate',
     'expirationDate',
     'lastUpdate',
@@ -161,12 +177,15 @@ export const schema = {
     'subtotal',
     'status',
     'seller',
+    'sellerName',
     'parentQuote',
     'hasChildren',
+    'childrenQuantity',
   ],
   'v-immediate-indexing': true,
   'v-indexed': [
     'creatorEmail',
+    'creatorName',
     'creationDate',
     'expirationDate',
     'lastUpdate',
@@ -176,6 +195,7 @@ export const schema = {
     'costCenter',
     'salesChannel',
     'seller',
+    'sellerName',
     'parentQuote',
     'hasChildren',
   ],