+++ node/typings.d.ts
@@ -2,6 +2,7 @@ interface Quote {
   id: string
   referenceName: string
   creatorEmail: string
+  creatorName: string
   creatorRole: string
   creationDate: string
   expirationDate: string
@@ -16,8 +17,10 @@ interface Quote {
   viewedByCustomer: boolean
   salesChannel: string | null
   seller?: string | null
+  sellerName?: string | null
   parentQuote?: string | null
   hasChildren?: boolean | null
+  childrenQuantity?: number | null
 }
 
 interface QuoteUpdate {
@@ -111,9 +114,10 @@ interface Settings {
     cronWorkspace?: string
     quotesManagedBy?: string
   }
-  hasSplittingQuoteFieldsInSchema?: boolean
   schemaVersion: string
+  schemaHash: string | null
   templateHash: string | null
+  storeName?: string | null
 }
 
 interface SessionData {
@@ -121,6 +125,8 @@ interface SessionData {
     profile: {
       id: { value: string }
       email: { value: string }
+      firstName: { value: string }
+      lastName: { value: string }
     }
     account: {
       accountName: { value: string }
@@ -135,12 +141,14 @@ interface SessionData {
 interface SellerQuoteInput {
   items: QuoteItem[]
   subtotal: number
+  sellerName: string
 }
 
 type SellerQuoteMap = Record<string, SellerQuoteInput>
 
 interface VerifyQuoteSettingsResponse {
   receiveQuotes: boolean
+  sellerName: string
 }
 
 interface SellerQuoteNotifyInput {